moved device details to root components folder
